% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chargement_Command_window.R
\name{Chargement_Command_window}
\alias{Chargement_Command_window}
\title{Easy progress bar}
\usage{
Chargement_command_window(size, text = '', minimal = FALSE)
}
\arguments{
\item{size}{Number of total elements in the loop}

\item{text}{Text to add before the progress bar (default = '')}

\item{minimal}{if FALSE, a progress bar will appear, if TRUE only the percentage is shown}
}
\description{
Progress bar easy to use to check progress of loops
}
\examples{
 
#simple loop
Chargement_Command_window(100)
for(i in 1:100)
 {
   Chargement_Command_window()
   Sys.sleep(0.15)
 }

#Use in double loops without progress bar with a descriptive text
Chargement_Command_window(6*7, text = 'Waiting...', minimal = T)
for (j in 1:6)
{
 for(i in 1:7)
 {
   Chargement_Command_window()
   Sys.sleep(0.15)
 }
}
}
