% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_read_time.R
\name{nc_read_time}
\alias{nc_read_time}
\title{Read time variable in a netCDF File}
\usage{
nc_read_time(filename)
}
\arguments{
\item{netcdf_file}{Path to the netCDF file}
}
\value{
vector with all dates in POSIXc format
}
\description{
Will read the "time" variable in a netcdf file, and transform in a POSIXc vector using the units of the time variable
}
\examples{

time <- nc_read_time("path\to\netCDFfile")

}
