% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec2time.R
\name{sec2time}
\alias{sec2time}
\title{Convert secondes to human readable time}
\usage{
sec2time(secondes)
}
\arguments{
\item{secondes}{number of seconds to convert}
}
\value{
character in format "days hours minuts secondes"
}
\description{
Convert a number of seconds in the format "days, hours min sec" to be easly readable by humain
}
\examples{
sec2time(361)
  [1] "6m 1s"

sec2time(c(18956, 900163, 23, 637))
  [1] "5h 15m 56s"   "10 days, 10h 2m 43s"   "23s"   "10m 37s"  
}
