% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_toDate.R
\name{text_toDate}
\alias{text_toDate}
\title{Convert date string to POSIXct}
\usage{
text_toDate(date)
}
\arguments{
\item{date}{date to convert as a string, with format %Y-%m-%d or %Y-%m-%d %H:%M:%S}
}
\value{
date as POSIXct
}
\description{
Easy to use string date to POSIXct convertor, with string dates in format %Y-%m-%d or %Y-%m-%d %H:%M:%S.
Any other format will result return NA
Timezone set for the conversion is 'GMT'
}
\examples{
date_POSIXct <- text_toDate('2012-12-26')
  [1] "2012-12-26 GMT"

text_toDate(c('2012-12-26 12:00:00','1995-01-16 20:00:00'))
  [1] "2012-12-26 12:00:00 GMT" "1995-01-16 20:00:00 GMT"
}
